Crate r2d2

A library providing a generic connection pool.

Modules

config
Pool configuration.

Structs

Config
A struct specifying the runtime configuration of a pool.

GetTimeout
An error returned by Pool::get if it times out without retrieving a connection.

InitializationError
An error returned by Pool::new if it fails to initialize connections.

LoggingErrorHandler
An ErrorHandler which logs at the error level.

NoopErrorHandler
An ErrorHandler which does nothing.

Pool
A generic connection pool.

PooledConnection
A smart pointer wrapping a connection.

Traits

ConnectionManager
A trait which provides connection-specific functionality.

ErrorHandler
A trait which handles errors reported by the ConnectionManager.