Struct r2d2::config::Builder

pub struct Builder<C, E> { /* fields omitted */ }

A builder for Config.

See the documentation of Config for more details about the default value and meaning of the configuration parameters.

Methods

impl<C, E: Error> Builder<C, E>

fn new() -> Builder<C, E>

Constructs a new Builder.

Parameters are initialized with their default values.

fn pool_size(self, pool_size: u32) -> Builder<C, E>

Sets pool_size.

Panics

Panics if pool_size is 0.

fn min_idle(self, min_idle: Option<u32>) -> Builder<C, E>

Sets min_idle.

fn helper_threads(self, helper_threads: u32) -> Builder<C, E>

Sets helper_threads.

Ignored if thread_pool is Some.

Panics

Panics if helper_threads is 0.

fn thread_pool(
self,
thread_pool: Option<Arc<ScheduledThreadPool>>
) -> Builder<C, E>

Sets the thread_pool.

fn test_on_check_out(self, test_on_check_out: bool) -> Builder<C, E>

Sets test_on_check_out.

fn initialization_fail_fast(
self,
initialization_fail_fast: bool
) -> Builder<C, E>

Sets initialization_fail_fast.

fn max_lifetime(self, max_lifetime: Option<Duration>) -> Builder<C, E>

Sets max_lifetime.

Panics

Panics if max_lifetime is the zero Duration.

fn idle_timeout(self, idle_timeout: Option<Duration>) -> Builder<C, E>

Sets idle_timeout.

Panics

Panics if idle_timeout is the zero Duration.

fn connection_timeout(self, connection_timeout: Duration) -> Builder<C, E>

Sets connection_timeout to the specified duration.

Panics

Panics if connection_timeout is the zero duration

fn error_handler(self, error_handler: Box<HandleError<E>>) -> Builder<C, E>

Sets the error_handler.

fn connection_customizer(
self,
connection_customizer: Box<CustomizeConnection<C, E>>
) -> Builder<C, E>

Sets the connection_customizer.

fn build(self) -> Config<C, E>

Consumes the Builder, turning it into a Config.

Panics

Panics if min_idle is larger than pool_size.