Struct r2d2::Extensions

pub struct Extensions(_);

A “type map” used to associate data with pooled connections.

Extensions is a data structure mapping types to a value of that type. This can be used to, for example, cache prepared statements along side their connection.

Implementations

impl Extensions

pub fn new() -> Extensions

Returns a new, empty Extensions.

pub fn insert<T>(&mut self, value: T) -> Option<T>where T: 'static + Sync + Send,

Inserts a new value into the map.

Returns the previously stored value of that type, if present.

pub fn get<T>(&self) -> Option<&T>where T: 'static + Sync + Send,

Returns a shared reference to the stored value of the specified type.

pub fn get_mut<T>(&mut self) -> Option<&mut T>where T: 'static + Sync + Send,

Returns a mutable reference to the stored value of the specified type.

pub fn remove<T>(&mut self) -> Option<T>where T: 'static + Sync + Send,

Removes the value of the specified type from the map, returning it.

pub fn clear(&mut self)

Removes all values from the map.