Struct r2d2::PooledConnection

pub struct PooledConnection<M> where
    M: ManageConnection,  { /* fields omitted */ }

A smart pointer wrapping a connection.

Implementations

impl<M> PooledConnection<M> where
M: ManageConnection,

pub fn extensions(this: &Self) -> &Extensions

Returns a shared reference to the extensions associated with this connection.

pub fn extensions_mut(this: &mut Self) -> &mut Extensions

Returns a mutable reference to the extensions associated with this connection.