Crate r2d2

A generic connection pool.

Opening a new database connection every time one is needed is both inefficient and can lead to resource exhaustion under high traffic conditions. A connection pool maintains a set of open connections to a database, handing them out for repeated use.

r2d2 is agnostic to the connection type it is managing. Implementors of the ManageConnection trait provide the database-specific logic to create and check the health of connections.

Example

Using an imaginary "foodb" database.

use std::thread;

extern crate r2d2;
extern crate r2d2_foodb;

fn main() {
    let config = r2d2::Config::builder()
        .pool_size(15)
        .error_handler(Box::new(r2d2::LoggingErrorHandler))
        .build();
    let manager = r2d2_foodb::FooConnectionManager::new("localhost:1234");

    let pool = r2d2::Pool::new(config, manager).unwrap();

    for _ in 0..20 {
        let pool = pool.clone();
        thread::spawn(move || {
            let conn = pool.get().unwrap();
            // use the connection
            // it will be returned to the pool when it falls out of scope.
        })
    }
}

Modules

config
Pool configuration.

Structs

Config
A struct specifying the runtime configuration of a pool.

GetTimeout
An error returned by Pool::get if it times out without retrieving a connection.

InitializationError
An error returned by Pool::new if it fails to initialize connections.

LoggingErrorHandler
A HandleError implementation which logs at the error level.

NopConnectionCustomizer
A CustomizeConnection which does nothing.

NopErrorHandler
A HandleError implementation which does nothing.

Pool
A generic connection pool.

PooledConnection
A smart pointer wrapping a connection.

Traits

CustomizeConnection
A trait which allows for customization of connections.

HandleError
A trait which handles errors reported by the ManageConnection.

ManageConnection
A trait which provides connection-specific functionality.