Trait r2d2::event::HandleEvent

pub trait HandleEvent: Debug + Sync + Send {
    fn handle_acquire(&self, event: AcquireEvent) { ... }

    fn handle_release(&self, event: ReleaseEvent) { ... }

    fn handle_checkout(&self, event: CheckoutEvent) { ... }

    fn handle_timeout(&self, event: TimeoutEvent) { ... }

    fn handle_checkin(&self, event: CheckinEvent) { ... }
}

A trait which is provided with information about events in a connection pool.

Provided methods

fn handle_acquire(&self, event: AcquireEvent)

Called when a new connection is acquired.

The default implementation does nothing.

fn handle_release(&self, event: ReleaseEvent)

Called when a connection is released.

The default implementation does nothing.

fn handle_checkout(&self, event: CheckoutEvent)

Called when a connection is checked out from the pool.

The default implementation does nothing.

fn handle_timeout(&self, event: TimeoutEvent)

Called when a checkout attempt times out.

The default implementation does nothing.

fn handle_checkin(&self, event: CheckinEvent)

Called when a connection is checked back into the pool.