Struct r2d2::Builder

pub struct Builder<M> where
    M: ManageConnection,  { /* fields omitted */ }

A builder for a connection pool.

Methods

impl<M> Builder<M> where
M: ManageConnection,

pub fn new() -> Builder<M>

Constructs a new Builder.

Parameters are initialized with their default values.

pub fn max_size(self, max_size: u32) -> Builder<M>

Sets the maximum number of connections managed by the pool.

Defaults to 10.

Panics

Panics if max_size is 0.

pub fn min_idle(self, min_idle: Option<u32>) -> Builder<M>

Sets the minimum idle connection count maintained by the pool.

If set, the pool will try to maintain at least this many idle connections at all times, while respecting the value of max_size.

Defaults to None (equivalent to the value of max_size).

pub fn thread_pool(self, thread_pool: Arc<ScheduledThreadPool>) -> Builder<M>

Sets the thread pool used for asynchronous operations such as connection creation.

Defaults to a new pool with 3 threads.

pub fn test_on_check_out(self, test_on_check_out: bool) -> Builder<M>

If true, the health of a connection will be verified via a call to ConnectionManager::is_valid before it is checked out of the pool.

Defaults to true.

pub fn max_lifetime(self, max_lifetime: Option<Duration>) -> Builder<M>

Sets the maximum lifetime of connections in the pool.

If set, connections will be closed after existing for at most 30 seconds beyond this duration.

If a connection reaches its maximum lifetime while checked out it will be closed when it is returned to the pool.

Defaults to 30 minutes.

Panics

Panics if max_lifetime is the zero Duration.

pub fn idle_timeout(self, idle_timeout: Option<Duration>) -> Builder<M>

Sets the idle timeout used by the pool.

If set, connections will be closed after sitting idle for at most 30 seconds beyond this duration.

Defaults to 10 minutes.

Panics

Panics if idle_timeout is the zero Duration.

pub fn connection_timeout(self, connection_timeout: Duration) -> Builder<M>

Sets the connection timeout used by the pool.

Calls to Pool::get will wait this long for a connection to become available before returning an error.

Defaults to 30 seconds.

Panics

Panics if connection_timeout is the zero duration

pub fn error_handler(
self,
error_handler: Box<HandleError<M::Error>>
) -> Builder<M>

Sets the handler for errors reported in the pool.

Defaults to the LoggingErrorHandler.

pub fn connection_customizer(
self,
connection_customizer: Box<CustomizeConnection<M::Connection, M::Error>>
) -> Builder<M>

Sets the connection customizer used by the pool.

Defaults to the NopConnectionCustomizer.

pub fn build(self, manager: M) -> Result<Pool<M>, Error>

Consumes the builder, returning a new, initialized pool.

It will block until the pool has established its configured minimum number of connections, or it times out.

Errors

Returns an error if the pool is unable to open its minimum number of connections.

Panics

Panics if min_idle is greater than max_size.

pub fn build_unchecked(self, manager: M) -> Pool<M>

Consumes the builder, returning a new pool.

Unlike build, this method does not wait for any connections to be established before returning.

Panics

Panics if min_idle is greater than max_size.