Struct r2d2::Pool

pub struct Pool<M: ManageConnection>(_);

A generic connection pool.

Methods

impl<M> Pool<M> where
M: ManageConnection,

pub fn new(manager: M) -> Result<Pool<M>, Error>

Creates a new connection pool with a default configuration.

pub fn builder() -> Builder<M>

Returns a builder type to configure a new pool.

pub fn get(&self) -> Result<PooledConnection<M>, Error>

Retrieves a connection from the pool.

Waits for at most the configured connection timeout before returning an error.

pub fn try_get(&self) -> Option<PooledConnection<M>>

Attempts to retrieve a connection from the pool if there is one available.

Returns None if there are no idle connections available in the pool. This method will not block waiting to establish a new connection.

pub fn state(&self) -> State

Returns information about the current state of the pool.

pub fn max_size(&self) -> u32

Returns the configured maximum pool size.

pub fn min_idle(&self) -> Option<u32>

Returns the configured mimimum idle connection count.

pub fn test_on_check_out(&self) -> bool

Returns if the pool is configured to test connections on check out.

pub fn max_lifetime(&self) -> Option<Duration>

Returns the configured maximum connection lifetime.

pub fn idle_timeout(&self) -> Option<Duration>

Returns the configured idle connection timeout.

pub fn connection_timeout(&self) -> Duration

Returns the configured connection timeout.