Struct r2d2::Pool

pub struct Pool<M: ManageConnection>(_);

A generic connection pool.

Methods

impl<M> Pool<M> where
M: ManageConnection,

fn new(
config: Config<M::Connection, M::Error>,
manager: M
) -> Result<Pool<M>, InitializationError>

Creates a new connection pool.

Returns an Err value if initialization_fail_fast is set to true in the configuration and the pool is unable to open all of its connections.

fn state(&self) -> State

Returns information about the current state of the pool.

fn config(&self) -> &Config<M::Connection, M::Error>

Returns the pool's configuration.

fn get(&self) -> Result<PooledConnection<M>, GetTimeout>

Retrieves a connection from the pool.

Waits for at most Config::connection_timeout before returning an error.

fn try_get(&self) -> Option<PooledConnection<M>>

Attempts to retrieve a connection from the pool if there is one available.

Returns None if there are no idle connections available in the pool. This method will not block waiting to establish a new connection.